# [None][feat] Multi-block mode for Hopper spec dec XQA kernel