@jhaotingc 2025-05-18
/bot run --disable-fail-fast