@jhaotingc 2025-05-19
/bot kill --post-merge