@tensorrt-cicd 2025-05-19
[PR_Github #5637](https://nv/trt-llm-cicd/job/helpers/job/PR_Github/5637/) Bot args parsing error: usage: /bot [-h]
            {run,kill,skip,submit,reviewers,reuse-pipeline,reuse-review} ...
/bot: error: unrecognized arguments: --post-merge