@lowsfer 2025-05-19
Good work! [gitlab/ftp/xqa](https://gitlab-master.nvidia.com/ftp/xqa) content is already opensource in trtllm repo, so please update the link.