@jhaotingc 2025-08-02
/bot run --disable-fail-fast